Uses NGINX and Boomi based ES
Uses NGINX as a reverse proxy for all HTTP connections as the training account has a limit of number of connections. So we route all HTTP via reverse proxy so Boomi sees the same URL and routing is done by NGNIX based on resource path set in Boomi. Also we are using a process based method to listen for KAFKA and then use HTTP connector to route those to ElasticSearch as opposed to v1, where we use Confluent's own ES connector.